Move kernel metadata gen into own test file
